sharing health check step with agent-skill program
