+++ src/lib/programs/events-audit/steps.ts
@@ -11,12 +11,7 @@
 import type { ProgramStep } from '@lib/programs/program-step';
 import type { WizardSession } from '@lib/wizard-session';
 import { RunPhase } from '@lib/wizard-session';
-import {
-  evaluateWizardReadiness,
-  WizardReadiness,
-  SIGNUP_WIZARD_READINESS_CONFIG,
-  getBlockingServiceKeys,
-} from '@lib/health-checks/readiness';
+import { HEALTH_CHECK_STEP } from '@lib/programs/shared/health-check-step';
 
 function needsSetup(session: WizardSession): boolean {
   const config = session.frameworkConfig;
@@ -27,27 +22,6 @@ function needsSetup(session: WizardSession): boolean {
   );
 }
 
-function healthCheckReady(session: WizardSession): boolean {
-  if (!session.readinessResult) return false;
-
-  if (session.signup) {
-    const hardBlocking = getBlockingServiceKeys(
-      session.readinessResult.health,
-      SIGNUP_WIZARD_READINESS_CONFIG,
-    );
-    const defaultBlocking = getBlockingServiceKeys(
-      session.readinessResult.health,
-    );
-    if (hardBlocking.length === 0 && defaultBlocking.length === 0) return true;
-    return session.outageDismissed;
-  }
-
-  if (session.readinessResult.decision === WizardReadiness.No) {
-    return session.outageDismissed;
-  }
-  return true;
-}
-
 export const EVENTS_AUDIT_PROGRAM: ProgramStep[] = [
   {
     id: 'intro',
@@ -55,25 +29,7 @@ export const EVENTS_AUDIT_PROGRAM: ProgramStep[] = [
     screenId: 'audit-intro',
     gate: (session) => session.setupConfirmed,
   },
-  {
-    id: 'health-check',
-    label: 'Health check',
-    screenId: 'health-check',
-    gate: healthCheckReady,
-    onInit: (ctx) => {
-      evaluateWizardReadiness()
-        .then((readiness) => {
-          ctx.setReadinessResult(readiness);
-        })
-        .catch(() => {
-          ctx.setReadinessResult({
-            decision: WizardReadiness.Yes,
-            health: {} as never,
-            reasons: [],
-          });
-        });
-    },
-  },
+  HEALTH_CHECK_STEP,
   {
     id: 'setup',
     label: 'Setup',
