+++ src/lib/programs/posthog-integration/steps.ts
@@ -9,12 +9,7 @@
 import type { ProgramStep } from '@lib/programs/program-step';
 import type { WizardSession } from '@lib/wizard-session';
 import { RunPhase } from '@lib/wizard-session';
-import {
-  evaluateWizardReadiness,
-  WizardReadiness,
-  SIGNUP_WIZARD_READINESS_CONFIG,
-  getBlockingServiceKeys,
-} from '@lib/health-checks/readiness';
+import { HEALTH_CHECK_STEP } from '@lib/programs/shared/health-check-step';
 import { detectPostHogIntegration } from './detect.js';
 
 function needsSetup(session: WizardSession): boolean {
@@ -24,27 +19,6 @@ function needsSetup(session: WizardSession): boolean {
   return config.metadata.setup.questions.some(
     (q: { key: string }) => !(q.key in session.frameworkContext),
   );
-}
-
-function healthCheckReady(session: WizardSession): boolean {
-  if (!session.readinessResult) return false;
-
-  if (session.signup) {
-    const hardBlocking = getBlockingServiceKeys(
-      session.readinessResult.health,
-      SIGNUP_WIZARD_READINESS_CONFIG,
-    );
-    const defaultBlocking = getBlockingServiceKeys(
-      session.readinessResult.health,
-    );
-    if (hardBlocking.length === 0 && defaultBlocking.length === 0) return true;
-    return session.outageDismissed;
-  }
-
-  if (session.readinessResult.decision === WizardReadiness.No) {
-    return session.outageDismissed;
-  }
-  return true;
 }
 
 export const POSTHOG_INTEGRATION_PROGRAM: ProgramStep[] = [
@@ -63,25 +37,7 @@ export const POSTHOG_INTEGRATION_PROGRAM: ProgramStep[] = [
     screenId: 'intro',
     gate: (session) => session.setupConfirmed,
   },
-  {
-    id: 'health-check',
-    label: 'Health check',
-    screenId: 'health-check',
-    gate: healthCheckReady,
-    onInit: (ctx) => {
-      evaluateWizardReadiness()
-        .then((readiness) => {
-          ctx.setReadinessResult(readiness);
-        })
-        .catch(() => {
-          ctx.setReadinessResult({
-            decision: WizardReadiness.Yes,
-            health: {} as never,
-            reasons: [],
-          });
-        });
-    },
-  },
+  HEALTH_CHECK_STEP,
   {
     id: 'setup',
     label: 'Setup',
